refactor: move PathManager (#4638)
