+++ engine-tests/src/main/java/org/terasology/engine/ModuleEnvironmentTest.java
@@ -5,9 +5,9 @@ package org.terasology.engine;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.io.TempDir;
+import org.terasology.engine.core.PathManager;
 import org.terasology.engine.core.module.ExternalApiWhitelist;
 import org.terasology.engine.core.module.ModuleManager;
-import org.terasology.engine.core.paths.PathManager;
 import org.terasology.engine.testUtil.ModuleManagerFactory;
 import org.terasology.reflection.ModuleTypeRegistry;
 import org.terasology.reflection.TypeRegistry;
