+++ engine-tests/src/test/java/org/terasology/engine/persistence/internal/StorageManagerTest.java
@@ -16,18 +16,12 @@ import org.junit.jupiter.api.TestMethodOrder;
 import org.junit.jupiter.api.io.TempDir;
 import org.mockito.ArgumentMatchers;
 import org.terasology.engine.TerasologyTestingEnvironment;
-import org.terasology.gestalt.assets.ResourceUrn;
-import org.terasology.gestalt.assets.management.AssetManager;
+import org.terasology.engine.core.PathManager;
 import org.terasology.engine.core.bootstrap.EntitySystemSetupUtil;
 import org.terasology.engine.core.module.ModuleManager;
-import org.terasology.engine.core.paths.PathManager;
 import org.terasology.engine.entitySystem.entity.EntityRef;
 import org.terasology.engine.entitySystem.entity.internal.EngineEntityManager;
-import org.terasology.unittest.stubs.EntityRefComponent;
-import org.terasology.unittest.stubs.StringComponent;
-import org.terasology.joml.geom.AABBfc;
 import org.terasology.engine.logic.location.LocationComponent;
-import org.terasology.gestalt.module.ModuleEnvironment;
 import org.terasology.engine.network.Client;
 import org.terasology.engine.network.ClientComponent;
 import org.terasology.engine.network.NetworkMode;
@@ -41,7 +35,6 @@ import org.terasology.engine.recording.RecordAndReplayCurrentStatus;
 import org.terasology.engine.recording.RecordAndReplaySerializer;
 import org.terasology.engine.recording.RecordAndReplayUtils;
 import org.terasology.engine.recording.RecordedEventStore;
-import org.terasology.reflection.TypeRegistry;
 import org.terasology.engine.registry.CoreRegistry;
 import org.terasology.engine.world.WorldProvider;
 import org.terasology.engine.world.block.Block;
@@ -54,6 +47,13 @@ import org.terasology.engine.world.chunks.ChunkProvider;
 import org.terasology.engine.world.chunks.blockdata.ExtraBlockDataManager;
 import org.terasology.engine.world.chunks.internal.ChunkImpl;
 import org.terasology.engine.world.internal.WorldInfo;
+import org.terasology.gestalt.assets.ResourceUrn;
+import org.terasology.gestalt.assets.management.AssetManager;
+import org.terasology.gestalt.module.ModuleEnvironment;
+import org.terasology.joml.geom.AABBfc;
+import org.terasology.reflection.TypeRegistry;
+import org.terasology.unittest.stubs.EntityRefComponent;
+import org.terasology.unittest.stubs.StringComponent;
 
 import java.io.File;
 import java.io.IOException;
