+++ engine/src/main/java/org/terasology/engine/config/Config.java
@@ -13,22 +13,22 @@ import com.google.gson.JsonParser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.terasology.engine.context.Context;
+import org.terasology.engine.core.PathManager;
 import org.terasology.engine.core.SimpleUri;
 import org.terasology.engine.core.TerasologyConstants;
-import org.terasology.engine.core.paths.PathManager;
 import org.terasology.engine.core.subsystem.Resolution;
 import org.terasology.engine.entitySystem.Component;
-import org.terasology.gestalt.naming.Version;
-import org.terasology.gestalt.naming.gson.NameTypeAdapter;
-import org.terasology.gestalt.naming.gson.VersionTypeAdapter;
-import org.terasology.input.Input;
-import org.terasology.gestalt.module.sandbox.API;
-import org.terasology.gestalt.naming.Name;
 import org.terasology.engine.utilities.gson.CaseInsensitiveEnumTypeAdapterFactory;
 import org.terasology.engine.utilities.gson.InputHandler;
 import org.terasology.engine.utilities.gson.ResolutionHandler;
 import org.terasology.engine.utilities.gson.SetMultimapTypeAdapter;
 import org.terasology.engine.utilities.gson.UriTypeAdapterFactory;
+import org.terasology.gestalt.module.sandbox.API;
+import org.terasology.gestalt.naming.Name;
+import org.terasology.gestalt.naming.Version;
+import org.terasology.gestalt.naming.gson.NameTypeAdapter;
+import org.terasology.gestalt.naming.gson.VersionTypeAdapter;
+import org.terasology.input.Input;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
