+++ engine/src/main/java/org/terasology/engine/core/module/ModuleInstaller.java
@@ -4,7 +4,7 @@ package org.terasology.engine.core.module;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.terasology.engine.core.paths.PathManager;
+import org.terasology.engine.core.PathManager;
 import org.terasology.engine.utilities.download.MultiFileDownloader;
 import org.terasology.engine.utilities.download.MultiFileTransferProgressListener;
 import org.terasology.gestalt.module.Module;
