+++ engine/src/main/java/org/terasology/engine/logic/behavior/BehaviorSystem.java
@@ -5,8 +5,7 @@ package org.terasology.engine.logic.behavior;
 import com.google.common.collect.Lists;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.terasology.gestalt.assets.ResourceUrn;
-import org.terasology.engine.core.paths.PathManager;
+import org.terasology.engine.core.PathManager;
 import org.terasology.engine.entitySystem.entity.EntityManager;
 import org.terasology.engine.entitySystem.entity.EntityRef;
 import org.terasology.engine.entitySystem.entity.lifecycleEvents.OnActivatedComponent;
@@ -22,10 +21,11 @@ import org.terasology.engine.logic.behavior.asset.BehaviorTreeData;
 import org.terasology.engine.logic.behavior.asset.BehaviorTreeFormat;
 import org.terasology.engine.logic.behavior.core.Actor;
 import org.terasology.engine.logic.behavior.core.BehaviorNode;
-import org.terasology.gestalt.assets.management.AssetManager;
-import org.terasology.gestalt.naming.Name;
 import org.terasology.engine.registry.In;
 import org.terasology.engine.registry.Share;
+import org.terasology.gestalt.assets.ResourceUrn;
+import org.terasology.gestalt.assets.management.AssetManager;
+import org.terasology.gestalt.naming.Name;
 
 import java.io.FileOutputStream;
 import java.io.IOException;
