+++ engine/src/main/java/org/terasology/engine/network/internal/ClientConnectionHandler.java
@@ -11,9 +11,9 @@ import org.slf4j.LoggerFactory;
 import org.terasology.engine.config.Config;
 import org.terasology.engine.config.PlayerConfig;
 import org.terasology.engine.core.EngineTime;
+import org.terasology.engine.core.PathManager;
 import org.terasology.engine.core.Time;
 import org.terasology.engine.core.module.ModuleManager;
-import org.terasology.engine.core.paths.PathManager;
 import org.terasology.engine.network.JoinStatus;
 import org.terasology.engine.registry.CoreRegistry;
 import org.terasology.gestalt.naming.Name;
