+++ engine/src/main/java/org/terasology/engine/persistence/internal/SaveTransaction.java
@@ -9,14 +9,13 @@ import org.joml.Vector3f;
 import org.joml.Vector3i;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.terasology.engine.core.paths.PathManager;
+import org.terasology.engine.core.PathManager;
 import org.terasology.engine.entitySystem.Component;
 import org.terasology.engine.entitySystem.entity.EntityRef;
 import org.terasology.engine.entitySystem.entity.internal.EngineEntityManager;
 import org.terasology.engine.game.GameManifest;
 import org.terasology.engine.logic.location.LocationComponent;
 import org.terasology.engine.network.ClientComponent;
-import org.terasology.protobuf.EntityData;
 import org.terasology.engine.recording.RecordAndReplayCurrentStatus;
 import org.terasology.engine.recording.RecordAndReplaySerializer;
 import org.terasology.engine.recording.RecordAndReplayStatus;
@@ -24,6 +23,7 @@ import org.terasology.engine.recording.RecordAndReplayUtils;
 import org.terasology.engine.utilities.concurrency.AbstractTask;
 import org.terasology.engine.world.chunks.Chunks;
 import org.terasology.engine.world.chunks.internal.ChunkImpl;
+import org.terasology.protobuf.EntityData;
 
 import java.io.BufferedOutputStream;
 import java.io.IOException;
