+++ engine/src/main/java/org/terasology/engine/rendering/nui/editor/layers/NUISkinEditorScreen.java
@@ -11,12 +11,20 @@ import com.google.gson.stream.JsonReader;
 import org.codehaus.plexus.util.ExceptionUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.terasology.gestalt.assets.ResourceUrn;
 import org.terasology.engine.config.Config;
 import org.terasology.engine.config.NUIEditorConfig;
-import org.terasology.engine.core.paths.PathManager;
+import org.terasology.engine.core.PathManager;
+import org.terasology.engine.registry.In;
+import org.terasology.engine.rendering.nui.asset.UIFormat;
 import org.terasology.engine.rendering.nui.editor.systems.NUISkinEditorSystem;
+import org.terasology.engine.rendering.nui.editor.utils.NUIEditorItemRenderer;
+import org.terasology.engine.rendering.nui.editor.utils.NUIEditorMenuTreeBuilder;
+import org.terasology.engine.rendering.nui.editor.utils.NUIEditorNodeUtils;
+import org.terasology.engine.rendering.nui.editor.utils.NUIEditorTextEntryBuilder;
+import org.terasology.engine.rendering.nui.skin.UISkinFormat;
 import org.terasology.engine.rendering.nui.widgets.JsonEditorTreeView;
+import org.terasology.engine.utilities.Assets;
+import org.terasology.gestalt.assets.ResourceUrn;
 import org.terasology.gestalt.assets.exceptions.InvalidUrnException;
 import org.terasology.gestalt.assets.format.AssetDataFile;
 import org.terasology.gestalt.assets.management.AssetManager;
@@ -37,14 +45,6 @@ import org.terasology.nui.widgets.UITextEntry;
 import org.terasology.nui.widgets.treeView.JsonTree;
 import org.terasology.nui.widgets.treeView.JsonTreeConverter;
 import org.terasology.nui.widgets.treeView.JsonTreeValue;
-import org.terasology.engine.registry.In;
-import org.terasology.engine.rendering.nui.asset.UIFormat;
-import org.terasology.engine.rendering.nui.editor.utils.NUIEditorItemRenderer;
-import org.terasology.engine.rendering.nui.editor.utils.NUIEditorMenuTreeBuilder;
-import org.terasology.engine.rendering.nui.editor.utils.NUIEditorNodeUtils;
-import org.terasology.engine.rendering.nui.editor.utils.NUIEditorTextEntryBuilder;
-import org.terasology.engine.rendering.nui.skin.UISkinFormat;
-import org.terasology.engine.utilities.Assets;
 
 import javax.swing.JDialog;
 import javax.swing.JFileChooser;
