+++ engine/src/main/java/org/terasology/engine/rendering/nui/layers/mainMenu/RecordScreen.java
@@ -4,16 +4,16 @@ package org.terasology.engine.rendering.nui.layers.mainMenu;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.terasology.gestalt.assets.ResourceUrn;
-import org.terasology.engine.core.paths.PathManager;
-import org.terasology.engine.rendering.nui.animation.MenuAnimationSystems;
-import org.terasology.engine.rendering.nui.layers.mainMenu.savedGames.GameInfo;
-import org.terasology.engine.rendering.nui.layers.mainMenu.savedGames.GameProvider;
-import org.terasology.nui.widgets.UIButton;
+import org.terasology.engine.core.PathManager;
 import org.terasology.engine.recording.RecordAndReplayCurrentStatus;
 import org.terasology.engine.recording.RecordAndReplayStatus;
 import org.terasology.engine.recording.RecordAndReplayUtils;
 import org.terasology.engine.registry.In;
+import org.terasology.engine.rendering.nui.animation.MenuAnimationSystems;
+import org.terasology.engine.rendering.nui.layers.mainMenu.savedGames.GameInfo;
+import org.terasology.engine.rendering.nui.layers.mainMenu.savedGames.GameProvider;
+import org.terasology.gestalt.assets.ResourceUrn;
+import org.terasology.nui.widgets.UIButton;
 
 import java.util.Objects;
 import java.util.stream.Stream;
