+++ engine/src/main/java/org/terasology/engine/rendering/nui/layers/mainMenu/SelectGameScreen.java
@@ -4,10 +4,9 @@ package org.terasology.engine.rendering.nui.layers.mainMenu;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.terasology.gestalt.assets.ResourceUrn;
 import org.terasology.engine.core.GameEngine;
+import org.terasology.engine.core.PathManager;
 import org.terasology.engine.core.modes.StateLoading;
-import org.terasology.engine.core.paths.PathManager;
 import org.terasology.engine.game.GameManifest;
 import org.terasology.engine.network.NetworkMode;
 import org.terasology.engine.registry.CoreRegistry;
@@ -15,6 +14,7 @@ import org.terasology.engine.rendering.nui.animation.MenuAnimationSystems;
 import org.terasology.engine.rendering.nui.layers.mainMenu.gameDetailsScreen.GameDetailsScreen;
 import org.terasology.engine.rendering.nui.layers.mainMenu.savedGames.GameInfo;
 import org.terasology.engine.rendering.nui.layers.mainMenu.savedGames.GameProvider;
+import org.terasology.gestalt.assets.ResourceUrn;
 import org.terasology.nui.databinding.ReadOnlyBinding;
 import org.terasology.nui.widgets.UIButton;
 import org.terasology.nui.widgets.UILabel;
