+++ engine/src/main/java/org/terasology/engine/rendering/nui/layers/mainMenu/savedGames/GameProvider.java
@@ -6,7 +6,7 @@ import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.terasology.engine.core.paths.PathManager;
+import org.terasology.engine.core.PathManager;
 import org.terasology.engine.game.GameManifest;
 
 import java.io.IOException;
