+++ engine/src/main/java/org/terasology/engine/world/block/tiles/WorldAtlasImpl.java
@@ -12,11 +12,7 @@ import gnu.trove.map.hash.TObjectIntHashMap;
 import org.joml.Vector2f;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.terasology.gestalt.assets.ResourceUrn;
-import org.terasology.engine.core.paths.PathManager;
-import org.terasology.joml.geom.Rectanglef;
-import org.terasology.math.TeraMath;
-import org.terasology.gestalt.naming.Name;
+import org.terasology.engine.core.PathManager;
 import org.terasology.engine.rendering.assets.atlas.Atlas;
 import org.terasology.engine.rendering.assets.atlas.AtlasData;
 import org.terasology.engine.rendering.assets.material.Material;
@@ -25,6 +21,10 @@ import org.terasology.engine.rendering.assets.texture.Texture;
 import org.terasology.engine.rendering.assets.texture.TextureData;
 import org.terasology.engine.rendering.assets.texture.subtexture.SubtextureData;
 import org.terasology.engine.utilities.Assets;
+import org.terasology.gestalt.assets.ResourceUrn;
+import org.terasology.gestalt.naming.Name;
+import org.terasology.joml.geom.Rectanglef;
+import org.terasology.math.TeraMath;
 
 import javax.imageio.ImageIO;
 import java.awt.Color;
