+++ facades/TeraEd/src/main/java/org/terasology/editor/TeraEd.java
@@ -1,4 +1,4 @@
-// Copyright 2020 The Terasology Foundation
+// Copyright 2021 The Terasology Foundation
 // SPDX-License-Identifier: Apache-2.0
 
 package org.terasology.editor;
@@ -12,10 +12,10 @@ import org.terasology.editor.subsystem.AwtInput;
 import org.terasology.editor.subsystem.LwjglPortlet;
 import org.terasology.editor.ui.MainWindow;
 import org.terasology.engine.core.GameEngine;
+import org.terasology.engine.core.PathManager;
 import org.terasology.engine.core.TerasologyEngine;
 import org.terasology.engine.core.TerasologyEngineBuilder;
 import org.terasology.engine.core.modes.StateMainMenu;
-import org.terasology.engine.core.paths.PathManager;
 import org.terasology.engine.core.subsystem.config.BindsSubsystem;
 import org.terasology.engine.core.subsystem.lwjgl.LwjglAudio;
 import org.terasology.engine.core.subsystem.lwjgl.LwjglTimer;
